Support servers-per-disk deployment
In the interest of supporting an upcoming Swift patch to facilitate
running one or more object-server processes per unique device port in
the rings (requires /having/ unique device ports in the rings), this
patch adds a new SERVERS_PER_PORT localrc setting (defaults to 0 which
is "off').  When on the setting is > 0:
 * Triggers object (storage policy) ring device ports to range between
   60x6 and 60x9, per disk, on each "server" (this limits
   DISKS <= 4 * NODES).  NOTE: replicated ring probe tests only work
   with exactly 4 disks, so you can't see much difference with this flag
   set, but you can on the EC ring (which has 8 disks and the probe
   tests work with *that*).
 * Enables the new "servers_per_port" flag in each object server's
   config.  NOTE: this must be done in the [DEFAULT] section so that
   the object-replicator and object-reconstructor can see it.
 * All servers get a bind_ip which is unique per "server"
   (127.0.0.1, 127.0.0.2, ...) and these IPs are configured in the rings
   as well.  NOTE: rsyncd must bind to 0.0.0.0 now, with the unique IPs.

With this new code, the vagrant SAIO still works with older swift, as
long as SERVER_PER_PORT is not enabled.

Unrelated addition: you can add an additional public key to
~vagrant/.ssh/authorized_keys in the localrc.

You can now allocate more CPUs to the SAIO using VAGRANT_CPUS in
localrc.

Bumped up default VM RAM size to 1 GiB (768 MB was insufficient to run
all services).